+++ source/F0.Mvvm/F0.Mvvm/Windows/Input/Command.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Threading.Tasks;
 
 namespace F0.Windows.Input
 {
@@ -22,6 +23,26 @@ namespace F0.Windows.Input
 		public static IInputCommand<T> Create<T>(Action<T> onExecute, Predicate<T> onCanExecute)
 		{
 			return new RelayCommand<T>(onExecute, onCanExecute);
+		}
+
+		public static IAsyncCommand Create(Func<Task> onExecute)
+		{
+			return new AsyncRelayCommand(onExecute);
+		}
+
+		public static IAsyncCommand Create(Func<Task> onExecute, Func<bool> onCanExecute)
+		{
+			return new AsyncRelayCommand(onExecute, onCanExecute);
+		}
+
+		public static IAsyncCommand<T> Create<T>(Func<T, Task> onExecute)
+		{
+			return new AsyncRelayCommand<T>(onExecute);
+		}
+
+		public static IAsyncCommand<T> Create<T>(Func<T, Task> onExecute, Predicate<T> onCanExecute)
+		{
+			return new AsyncRelayCommand<T>(onExecute, onCanExecute);
 		}
 	}
 }